fix: STOR-47
