+++ src/components/Tabs/Tabs.scss
