+++ src/components/Tabs/Tabs.stories.js
@@ -1,9 +1,9 @@
 import Tabs from './Tabs.vue';
 
 export default {
-    title: 'Display/Tabs',
-    component: Tabs,
-    parameters: {
+	title: 'Display/Tabs',
+	component: Tabs,
+	parameters: {
 		docs: {
 			description: {
 				component: `Tabs<br />
@@ -17,5 +17,52 @@ export default {
 };
 
 export const Primary = () => ({
-    template: '<farm-tabs />',
+	data() {
+		return {
+			v: 'Seleção',
+		};
+	},
+	methods: {
+		updateValue(value) {
+			this.v = value.name;
+		},
+	},
+	template: `<div>
+				<farm-tabs @update="updateValue" /> 
+				<p class="mt-2">selected: {{v}}</p>
+			</div>`,
+});
+
+export const WithoutCounter = () => ({
+	data() {
+		return {
+			v: 'Seleção',
+		};
+	},
+	methods: {
+		updateValue(value) {
+			this.v = value.name;
+		},
+	},
+	template: `<div>
+				<farm-tabs :showCounter="false" @update="updateValue" /> 
+				<p class="mt-2">selected: {{v}}</p>
+			</div>`,
+});
+
+export const Disabled = () => ({
+	data() {
+		return {
+			v: 'Seleção',
+		};
+	},
+	methods: {
+		updateValue(value) {
+			this.v = value.name;
+		},
+	},
+	template: `<div>
+				<farm-tabs :showCounter="false" :allow-user-change="false" @update="updateValue" /> 
+				<p class="mt-2">selected: {{v}}</p>
+			</div>`,
 });
