+++ src/components/Tabs/index.ts
@@ -1,3 +1,4 @@
 import Tabs from './Tabs.vue';
 
+export { Tabs };
 export default Tabs;
