--- src/components/TabsV2/index.ts
@@ -1,3 +1,0 @@
-import TabsV2 from './TabsV2.vue';
-
-export default TabsV2;
