+++ src/main.ts
@@ -6,8 +6,6 @@ import MainFilter from './components/MainFilter';
 import Loader from './components/Loader';
 import FilePicker from './components/FilePicker';
 import MultipleFilePicker from './components/MultipleFilePicker';
-import Tabs from './components/Tabs';
-import TabsV2 from './components/TabsV2';
 import DialogHeader from './components/DialogHeader';
 import DialogFooter from './components/DialogFooter';
 import RangeDatePicker from './components/RangeDatePicker';
@@ -33,8 +31,6 @@ export {
 	Loader,
 	FilePicker,
 	MultipleFilePicker,
-	Tabs,
-	TabsV2,
 	DialogHeader,
 	DialogFooter,
 	RangeDatePicker,
@@ -87,6 +83,7 @@ export * from './components/RadioGroup';
 export * from './components/Select';
 export * from './components/Stepper';
 export * from './components/Switcher';
+export * from './components/Tabs';
 export * from './components/TextArea';
 export * from './components/TextField';
 export * from './components/TextFieldV2';
